// studio-web/src/api/models/campaignDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: v1
 */
import type { EmailTemplateDTO } from './emailTemplateDTO';
import type { CampaignRecipientDTO } from './campaignRecipientDTO';

export interface CampaignDTO {
  /** @nullable */
  id?: string | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @minLength 1 */
  emailTemplateId: string;
  emailTemplate?: EmailTemplateDTO;
  /** @minLength 1 */
  status: string;
  totalRecipients?: number;
  sentCount?: number;
  openCount?: number;
  errorCount?: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt?: string;
  /** @minLength 1 */
  createdBy: string;
  /** @nullable */
  recipients?: CampaignRecipientDTO[] | null;
}